fix: bump series id from uint8 to uint32

+++ src/AlephVaultBase.sol
@@ -97,9 +97,9 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
         view
         returns (uint256)
     {
-        uint8 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
+        uint32 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
         uint256 _totalAssetsSum;
-        for (uint8 _seriesId; _seriesId <= _shareClass.shareSeriesId; _seriesId++) {
+        for (uint32 _seriesId; _seriesId <= _shareClass.shareSeriesId; _seriesId++) {
             if (_seriesId > SeriesAccounting.LEAD_SERIES_ID) {
                 _seriesId += _lastConsolidatedSeriesId;
             }
@@ -116,7 +116,7 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
      * @param _seriesId The ID of the share series.
      * @return The total assets in the vault.
      */
-    function _totalAssetsPerSeries(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint8 _seriesId)
+    function _totalAssetsPerSeries(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint32 _seriesId)
         internal
         view
         returns (uint256)
@@ -131,7 +131,7 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
      * @param _seriesId The ID of the share series.
      * @return The total shares in the vault.
      */
-    function _totalSharesPerSeries(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint8 _seriesId)
+    function _totalSharesPerSeries(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint32 _seriesId)
         internal
         view
         returns (uint256)
@@ -146,7 +146,7 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
      * @param _user The user to get the shares of.
      * @return The shares of the user.
      */
-    function _sharesOf(IAlephVault.ShareClass storage _shareClass, uint8 _seriesId, address _user)
+    function _sharesOf(IAlephVault.ShareClass storage _shareClass, uint32 _seriesId, address _user)
         internal
         view
         returns (uint256)
@@ -162,7 +162,7 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
      * @param _user The user to get the assets of.
      * @return The assets of the user.
      */
-    function _assetsOf(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint8 _seriesId, address _user)
+    function _assetsOf(IAlephVault.ShareClass storage _shareClass, uint8 _classId, uint32 _seriesId, address _user)
         internal
         view
         returns (uint256)
@@ -186,9 +186,9 @@ contract AlephVaultBase is ReentrancyGuardUpgradeable {
         returns (uint256)
     {
         uint256 _assets;
-        uint8 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
-        uint8 _shareSeriesId = _shareClass.shareSeriesId;
-        for (uint8 _seriesId; _seriesId <= _shareSeriesId; _seriesId++) {
+        uint32 _lastConsolidatedSeriesId = _shareClass.lastConsolidatedSeriesId;
+        uint32 _shareSeriesId = _shareClass.shareSeriesId;
+        for (uint32 _seriesId; _seriesId <= _shareSeriesId; _seriesId++) {
             if (_seriesId > SeriesAccounting.LEAD_SERIES_ID) {
                 _seriesId += _lastConsolidatedSeriesId;
             }
